Fix bug
When drawing dots in the signature pad, it will be considered as empty (signature.isEmpty) because mouseup event was handled in angular-signature before being handled in signature_pad.js (handling the event in signature_pad.js will set this._isEmpty to false).
However, this is not the case when drawing a curve in the signature pad, because this._isEmpty will be set to false when handling the mousemove event which will be fired before the mouseup event. So when the latter is fired, this._isEmpty was already set to false and the signature pad won't be considered as empty.
The solution is to defer handling mouseup event in angular-signature (ensuring that handling the same event in signature_pad.js will be done first)